refactor(paw): split cloudflared service state
